chore: add max lines linting rules and refactor
