+++ backend/core/archer.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import {
   AmiChart,
   EnumJurisdictionLanguages,
