+++ backend/core/src/listings/entities/listing.entity.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import {
   BaseEntity,
   Column,
