+++ ui-components/src/helpers/preferences.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import React from "react"
 import {
   InputType,
