+++ ui-components/src/icons/Icons.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import * as React from "react"
 
 export interface IconProps {
